refactor: constant 폴더 정리
