+++ src/components/@common/ImageViewer/index.tsx
@@ -8,7 +8,7 @@ import 'swiper/css/navigation';
 import 'swiper/css/pagination';
 import { useEffect, useRef, useState } from 'react';
 import useLockBodyScroll from '@/hooks/@common/useLockBodyScroll';
-import PageNavigator from '@/components/home/PageNavigator';
+import PageNavigator from '@/components/@common/PageNavigator';
 
 interface IImageViewer {
   /** 표시할 이미지 목록 (이미지 경로 배열) */
