+++ src/components/@common/PageNavigator/PageNavigator.styled.ts
