+++ src/components/home/HomeFriendListByTag/index.tsx
@@ -10,10 +10,9 @@ import { getRandomNumbers } from '@/utils/random';
 const HomeFriendListByTag = () => {
   const [categoryIds, setCategoryIds] = useState<number[]>([]);
 
-  // 전역 변수에 저장된 사용자 정보 가져오기
   const userInfo = useRecoilValue(profileAtom);
 
-  // 사용자가 선택한 관심사 배열 원소 중, 3개만 추출해서 세팅
+  // 사용자가 선택한 관심사 배열 원소 중, 3개만 추출
   useEffect(() => {
     if (userInfo) {
       const userSelectedTag = userInfo.interestTag
