+++ src/constants/user/month/index.ts
@@ -1,7 +1,10 @@
 import { createMonthOption } from '@/utils/date';
 
-const MONTH_COUNT = 12;
+export const MONTH_CONSTANT = Object.freeze({
+  MONTH_COUNT: 12,
+});
 
-export const BIRTH_MONTH = Array.from({ length: MONTH_COUNT }, (_, i) =>
-  createMonthOption(i),
+export const BIRTH_MONTH = Array.from(
+  { length: MONTH_CONSTANT.MONTH_COUNT },
+  (_, i) => createMonthOption(i),
 );
